+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.19.0.0
 
-* Export `dRepToText`
+* Add field accessors to `Block`: `blockHeader` and `blockBody`.
+* Expose `dRepToText`
 * Modify `withdrawalsThatDoNotDrainAccounts` to return `Maybe (Withdrawals, Withdrawals)` where the `fst` are either missing accounts or in the wrong network and `snd` are incomplete withdrawals.
 * Add `FromJSON` instance for `PParamUpdate`